# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SERMACROPS EDI Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /partners:
    get:
      operationId: listPartners
      tags: [partners]
      summary: List all trading partners
      responses:
        "200":
          description: List of trading partners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradingPartner"
    post:
      operationId: createPartner
      tags: [partners]
      summary: Create a trading partner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradingPartnerInput"
      responses:
        "201":
          description: Created partner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradingPartner"
  /partners/{id}:
    get:
      operationId: getPartner
      tags: [partners]
      summary: Get a trading partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trading partner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradingPartner"
        "404":
          description: Not found
    patch:
      operationId: updatePartner
      tags: [partners]
      summary: Update a trading partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradingPartnerUpdate"
      responses:
        "200":
          description: Updated partner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradingPartner"
    delete:
      operationId: deletePartner
      tags: [partners]
      summary: Delete a trading partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── Partner Endpoints ────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    TradingPartner:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        as2Identifier:
          type: string
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        supportedDocTypes:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, pending]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - as2Identifier
        - supportedDocTypes
        - status
        - createdAt
        - updatedAt
    TradingPartnerInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        as2Identifier:
          type: string
          minLength: 1
        contactEmail:
          type: string
        contactPhone:
          type: string
        supportedDocTypes:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, pending]
      required:
        - name
        - as2Identifier
    TradingPartnerUpdate:
      type: object
      properties:
        name:
          type: string
        as2Identifier:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        supportedDocTypes:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, pending]

    # ─── Partner Endpoints ──────────────────────────────────────────────────────
...
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [status_change, validation_error, transmission_failure, acknowledgment, system]
        title:
          type: string
        message:
          type: string
        documentId:
          type: ["string", "null"]
        partnerId:
          type: ["string", "null"]
        isRead:
          type: boolean
        severity:
          type: string
          enum: [info, warning, error, success]
        createdAt:
          type: string
      required:
        - id
        - type
        - title
        - message
        - isRead
        - severity
        - createdAt

    # ─── AS2 Settings ───────────────────────────────────────────────────────────
...